block: posix: Handle undetectable alignment
We have some cases when bug_align or request_alignment cannot be
detected on storage with 4k sector size:

- buf_align when probing Gluster storage. In this case the aligned
  buffer in qemu is not used to do the actual I/O on the server side, so
  the alignment cannot be detected.

- buf_align when probing unallocated file on XFS file system. Looks
  like XFS optimize the I/O and succeed with unaligned buffer.

- request_alignment when probing empty or unallocated file on Gluster
  storage backed by XFS filesystem.

These cases work when storage sector size is 512 bytes, because the
current code starts checking align=512. If the check succeeds because
alignment cannot be detected we use 512.

Practically the alignment requirements are the same for buffer
alignment, buffer length, and offset in file. So in case we cannot
detect buf_align, we can use request alignment. If we cannot detect
request alignment, we can fallback to a safe value.

With this change, we can provision and copy disks on storage with 4k
in the case mentioned above. Since we fallback to a safe alignment, it
is possible that we will do some unneeded data copies when aligning
buffers, but it is better than failing.

Resolves:
- https://bugzilla.redhat.com/1737256
- https://bugzilla.redhat.com/1738657

Signed-off-by: Nir Soffer <nsoffer@redhat.com>